Файлы переписывал на ts, некоторые исправлял в угоду линтера
