+++ .eslintrc.json
@@ -26,7 +26,9 @@
     "linebreak-style": "off",
     "@typescript-eslint/no-var-requires": 0,
     "semi": "off",
-    "@typescript-eslint/semi": "error"
+    "@typescript-eslint/semi": "error",
+    "@typescript-eslint/no-use-before-define": "off",
+    "no-use-before-define": 0
   },
   "settings": {
     "react": {
