+++ src/types/Iimg.ts
@@ -1,0 +1,5 @@
+interface Iimg {
+  img: string;
+}
+
+export default Iimg;
