+++ src/types/IisOpen.ts
@@ -1,0 +1,5 @@
+interface IisOpen {
+  isOpen: boolean;
+}
+
+export default IisOpen;
